Fix handling of available variable scopes
It is not necessary, and in some cases it is not even valid, to
update the lists of available scope values before submitting the
modified library variable set.  Library variables cannot be scoped
to actions, so there is no need to update those.  Roles are
identified by the role name and do not have a server-specific
unique ID that needs to be mapped.  The environments and machines
are checked to make sure they exist before attempting to import
the updated library variable set, so all referenced values are
already guaranteed to exist.  If the list of environments is
modified and does not exactly match the current list of environments
on the target Octopus server then some versions of Octopus will fail
to update the variables and return an error indicating that the
environment has changed.  The same is probably true for machines as
well.  The end result is that the reference ids in the variable
definitions need to be mapped, but the existing lists of scope
values should not be modified.